+++ src/app/SignUp.tsx
@@ -1,0 +1,10 @@
+/** @format */
+
+import React from "react";
+import { Text } from "react-native";
+
+const SignUp = () => {
+  return <Text>SignUpページ</Text>;
+};
+
+export default SignUp;
