+++ src/app/collection/CollectionDetail.tsx
@@ -1,0 +1,10 @@
+/** @format */
+
+import React from "react";
+import { Text } from "react-native";
+
+const CollectionDetail = () => {
+  return <Text>CollectionDetailページ</Text>;
+};
+
+export default CollectionDetail;
