+++ src/app/index.tsx
@@ -1,10 +1,18 @@
 /** @format */
 
 import React from "react";
-import { Redirect } from "expo-router";
+import { Link, Redirect } from "expo-router";
+import { Text } from "react-native";
 
 const Index = () => {
-    return <Redirect href='question/nottori' />
-}
+  return (
+    <>
+      <Text>Indexページ</Text>
+      <Link href="./SignUp">新規登録ページに移動</Link>
+      <Link href="./Login">ログインページに移動</Link>
+      <Link href="./collection/">collectionページに移動</Link>
+    </>
+  );
+};
 
 export default Index;
